+++ src/errors/InvalidServiceAlias.ts
@@ -1,0 +1,6 @@
+export class InvalidServiceAlias extends Error {
+  constructor(serviceName: string) {
+    super(`[${serviceName}]: should start with "@"`);
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
