+++ src/errors/NoClassDeclared.ts
@@ -1,0 +1,6 @@
+export class NoClassDeclared extends Error {
+  constructor(serviceName: string) {
+    super(`[${serviceName}]: No class found for this file`);
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
