+++ src/errors/ParameterAlreadyDeclared.ts
@@ -1,0 +1,6 @@
+export class ParameterAlreadyDeclared extends Error {
+  constructor(parameterName: string) {
+    super(`[${parameterName}]: Parameter already declared`);
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
