+++ src/errors/ServiceNotExist.ts
@@ -1,0 +1,6 @@
+export class ServiceNotExist extends Error {
+  constructor(serviceName: string) {
+    super(`[${serviceName}]: Service does not exist`);
+    Object.setPrototypeOf(this, new.target.prototype);
+  }
+}
